Remove obsolete actions and filesystem comparers
This commit deletes outdated actions and filesystem comparison classes. It covers classes in the Zafiro.FileSystem namespaces wherein several modules relating to file and directory actions, as well as filesystem comparers, were found to be outdated and thus have been removed. This clean-up will enhance the readability and maintainability of the codebase by eradicating unnecessary clutter.